Refactored Code a bit for easier programming
Moved all Troop functions into Troop File, same with base functions.
Made a few minor edits to code to make it easier to edit.
